Add Tooltip.stories.tsx + rename folders
