+++ src/stories/nrn/animation/BubbleTimer.stories.tsx
@@ -6,7 +6,7 @@ import React from 'react';
 import BubbleTimer, { Props } from '../../../components/animations/BubbleTimer';
 
 export default {
-  title: 'Next Right Now/Animations/BubbleTimer',
+  title: 'Next Right Now/Animation/BubbleTimer',
   component: BubbleTimer,
   argTypes: {},
 } as Meta;
