+++ src/stories/sb-examples/Button.stories.tsx
@@ -5,7 +5,7 @@ import { Story, Meta } from '@storybook/react/types-6-0';
 import { Button, ButtonProps } from './Button';
 
 export default {
-  title: 'Storybook/Examples/Button',
+  title: 'Storybook/Default examples/Button',
   component: Button,
   argTypes: {
     backgroundColor: { control: 'color' },
